backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
dj-database-url
python-dotenv
gunicorn